+++ internal/orchestrator/additional_helpers_test.go
@@ -913,7 +913,12 @@ func TestExtractArchiveNativeSymlinkAndHardlink(t *testing.T) {
 	}
 
 	dest := filepath.Join(tmpDir, "dest")
-	if err := extractArchiveNative(context.Background(), tarPath, dest, logger, nil, RestoreModeFull, nil, "", nil); err != nil {
+	if err := extractArchiveNative(context.Background(), restoreArchiveOptions{
+		archivePath: tarPath,
+		destRoot:    dest,
+		logger:      logger,
+		mode:        RestoreModeFull,
+	}); err != nil {
 		t.Fatalf("extractArchiveNative error: %v", err)
 	}
 
@@ -1292,7 +1297,12 @@ func TestExtractArchiveNativeBlocksTraversal(t *testing.T) {
 	_ = f.Close()
 
 	dest := filepath.Join(tmpDir, "dest")
-	if err := extractArchiveNative(context.Background(), tarPath, dest, logger, nil, RestoreModeFull, nil, "", nil); err != nil {
+	if err := extractArchiveNative(context.Background(), restoreArchiveOptions{
+		archivePath: tarPath,
+		destRoot:    dest,
+		logger:      logger,
+		mode:        RestoreModeFull,
+	}); err != nil {
 		t.Fatalf("extractArchiveNative error: %v", err)
 	}
 	if _, err := os.Stat(filepath.Join(dest, "../etc/passwd")); err == nil {
